+++ app/src/main/kotlin/org/eclipse/kuksa/testapp/databroker/DataBrokerEngine.kt
@@ -20,11 +20,11 @@
 package org.eclipse.kuksa.testapp.databroker
 
 import android.content.Context
-import org.eclipse.kuksa.CoroutineCallback
-import org.eclipse.kuksa.DataBrokerConnection
-import org.eclipse.kuksa.DisconnectListener
-import org.eclipse.kuksa.PropertyListener
-import org.eclipse.kuksa.VssNodeListener
+import org.eclipse.kuksa.coroutine.CoroutineCallback
+import org.eclipse.kuksa.connectivity.databroker.DataBrokerConnection
+import org.eclipse.kuksa.connectivity.databroker.listener.DisconnectListener
+import org.eclipse.kuksa.connectivity.databroker.listener.PropertyListener
+import org.eclipse.kuksa.connectivity.databroker.listener.VssNodeListener
 import org.eclipse.kuksa.model.Property
 import org.eclipse.kuksa.proto.v1.KuksaValV1.GetResponse
 import org.eclipse.kuksa.proto.v1.KuksaValV1.SetResponse
