+++ kuksa-sdk/src/main/kotlin/org/eclipse/kuksa/connectivity/authentication/VALStubExtension.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,9 @@
  * limitations under the License.
  *
  * SPDX-License-Identifier: Apache-2.0
- *
  */
 
-package org.eclipse.kuksa.authentication
+package org.eclipse.kuksa.connectivity.authentication
 
 import com.google.common.net.HttpHeaders
 import io.grpc.ClientInterceptor
