+++ kuksa-sdk/src/main/kotlin/org/eclipse/kuksa/connectivity/databroker/listener/PropertyListener.kt
@@ -14,10 +14,9 @@
  * limitations under the License.
  *
  * SPDX-License-Identifier: Apache-2.0
- *
  */
 
-package org.eclipse.kuksa
+package org.eclipse.kuksa.connectivity.databroker.listener
 
 import org.eclipse.kuksa.pattern.listener.Listener
 import org.eclipse.kuksa.proto.v1.KuksaValV1
