+++ samples/src/main/java/com/example/sample/JavaActivity.java
@@ -22,13 +22,13 @@ package com.example.sample;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import org.eclipse.kuksa.CoroutineCallback;
-import org.eclipse.kuksa.DataBrokerConnection;
-import org.eclipse.kuksa.DataBrokerConnector;
-import org.eclipse.kuksa.DisconnectListener;
-import org.eclipse.kuksa.PropertyListener;
-import org.eclipse.kuksa.VssNodeListener;
-import org.eclipse.kuksa.authentication.JsonWebToken;
+import org.eclipse.kuksa.coroutine.CoroutineCallback;
+import org.eclipse.kuksa.connectivity.databroker.DataBrokerConnection;
+import org.eclipse.kuksa.connectivity.databroker.DataBrokerConnector;
+import org.eclipse.kuksa.connectivity.databroker.listener.DisconnectListener;
+import org.eclipse.kuksa.connectivity.databroker.listener.PropertyListener;
+import org.eclipse.kuksa.connectivity.databroker.listener.VssNodeListener;
+import org.eclipse.kuksa.connectivity.authentication.JsonWebToken;
 import org.eclipse.kuksa.model.Property;
 import org.eclipse.kuksa.proto.v1.KuksaValV1;
 import org.eclipse.kuksa.proto.v1.KuksaValV1.GetResponse;
