+++ samples/src/main/kotlin/com/example/sample/KotlinActivity.kt
@@ -26,13 +26,13 @@ import io.grpc.Grpc
 import io.grpc.ManagedChannelBuilder
 import io.grpc.TlsChannelCredentials
 import kotlinx.coroutines.launch
-import org.eclipse.kuksa.DataBrokerConnection
-import org.eclipse.kuksa.DataBrokerConnector
-import org.eclipse.kuksa.DataBrokerException
-import org.eclipse.kuksa.DisconnectListener
-import org.eclipse.kuksa.PropertyListener
-import org.eclipse.kuksa.VssNodeListener
-import org.eclipse.kuksa.authentication.JsonWebToken
+import org.eclipse.kuksa.connectivity.databroker.DataBrokerConnection
+import org.eclipse.kuksa.connectivity.databroker.DataBrokerConnector
+import org.eclipse.kuksa.connectivity.databroker.DataBrokerException
+import org.eclipse.kuksa.connectivity.databroker.listener.DisconnectListener
+import org.eclipse.kuksa.connectivity.databroker.listener.PropertyListener
+import org.eclipse.kuksa.connectivity.databroker.listener.VssNodeListener
+import org.eclipse.kuksa.connectivity.authentication.JsonWebToken
 import org.eclipse.kuksa.model.Property
 import org.eclipse.kuksa.proto.v1.KuksaValV1
 import org.eclipse.kuksa.proto.v1.Types
